added a flag in viewmodel isGameInProgress to check the game progress, created ui test to ramdom clicks on cell after game is finished.
Why:
While doing a random UI test in a slow emulator, after the game is won and the alert dialog is show the ui test was able to click the cell even though the alert dialog is shown. In order to pass this test case , the flag was added so that the cell will not be clicked after the dialog is shown. Once the dialog is dismissed the game state will be restored to normal state and the cells will be available for interactions.